Convert .properties files to .example files
Change .properties file extension to .properties.example and keep only these
files in the repo. .properties files should be created locally because they may
be different on different systems and environments so they should not be
in the repo.

This also allows to make local-runner repo as a submodule in other git repos
without modification any files.